Add onPacket to IPCListener & relay echoed packets
This commit adds an onPacket listener to IPCListener and adds support
for echoed packets from the client.

When we send a packet, the client will send that packet back once it
has been acknowledged successfully. In the future, we will wait for
this response before returning from sendPacket, but for now we will
just inform the listener of the packet.