Introduce the concept of 'inbound' and 'outbound' packets
In our previous packet system, we only had one type of packet that had
no sense of which direction it was going. This made it hard to parse the
data of inbound packets, as sometimes their data structure was different to
the outbound packet.

This commit removes the old packet system, along with those disgusting
packet handlers, and moves to an 'inbound' and 'outbound' packet system.

The next thing we need to do is introduce timeouts, and use nonces to track
a packet's lifetime.