Utilise the nonce field in the packet pipeline
Before, we were ignoring the nonce field, this meant that we couldn't
reliably track when we received a response from Discord after sending a packet.

Now, sendPacket<T>(packet) will take in the Inbound version of your packet
as a type parameter (T). It will then return an instance of your Inbound
packet to you once a packet which is of the same instance, and the same
nonce, has been received.

Right now, we don't have any sort of timeout for how long a packet send
should take, this means that if the Discord client closes unexpectedly,
or something goes wrong in our pipeline, the sendPacket function will wait
forever. This is something that needs to be addressed in the form of timeouts.