fix: read all packet data before continuing to decode
If a packet is above 8192 bytes in length, Discord sends it in multiple
chunks instead of just one. We treated it as if all data comes in at
once. This is fine for smaller packets, but for larger packets we need
to keep reading until we have the same amount of data as the length.

This fixes an issue when using the `GET_RELATIONSHIPS` command on Unix
systems. I'm unsure if this issue occurs on Windows or not.