Everywhere: Handle disconnections better
We now have a new custom event, called DisconnectedEvent. This will
allow you to listen to when the connection is closed forcefully, or via
`ipc.disconnect()`. Although, it will provide no extra information.

Previously, we were only catching SocketException when reading.
SocketException extends IOException, but RandomAccessFile isn't
exactly a socket, so it just throws raw IOExceptions when a connection
failure happens. So, we should just catch IOException and check the
content of the message to see what it's about.

We also now catch IOExceptions when writing too, so this should fix
#15.

This system could definitely be improved, but for now, this is OK.